+++ src/main/java/study/Car.java
@@ -4,8 +4,10 @@ public class Car {
 
     private int position;
 
-    public void moveCar() {
-        position++;
+    public void moveCar(CarMovingStrategy carMovingStrategy) {
+        if (carMovingStrategy.canMoveCar()) {
+            position++;
+        }
     }
 
     public int getPosition() {
