+++ src/main/java/study/CarMovingStrategy.java
@@ -1,0 +1,7 @@
+package study;
+
+public interface CarMovingStrategy {
+
+    boolean canMoveCar();
+
+}
